Estranged nephew of more truncation fixes.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@5265 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/OXPVerifier/OOTextureVerifierStage.m
@@ -164,8 +164,8 @@ static NSString * const kStageName	= @"Testing textures and images";
 		
 		if (success)
 		{
-			rWidth = OORoundUpToPowerOf2((2 * pixmap.width) / 3);
-			rHeight = OORoundUpToPowerOf2((2 * pixmap.height) / 3);
+			rWidth = (uint32_t)OORoundUpToPowerOf2((2 * pixmap.width) / 3);
+			rHeight = (uint32_t)OORoundUpToPowerOf2((2 * pixmap.height) / 3);
 			if (pixmap.width != rWidth || pixmap.height != rHeight)
 			{
 				OOLog(@"verifyOXP.texture.notPOT", @"----- WARNING: image %@ has non-power-of-two dimensions; it will have to be rescaled (from %ux%u pixels to %ux%u pixels) at runtime.", displayName, pixmap.width, pixmap.height, rWidth, rHeight);
